修改数据库部分字段的数据类型

+++ src/main/java/cn/waynechu/mmall/mapper/CartMapper.java
@@ -6,21 +6,21 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface CartMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Cart record);
 
     int insertSelective(Cart record);
 
-    Cart selectByPrimaryKey(Integer id);
+    Cart selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Cart record);
 
     int updateByPrimaryKey(Cart record);
 
-    Cart selectByUserIdAndProductId(@Param("userId") Integer userId, @Param("productId") Integer productId);
+    Cart selectByUserIdAndProductId(@Param("userId") Long userId, @Param("productId") Long productId);
 
-    List<Cart> selectByUserId(Integer userId);
+    List<Cart> selectByUserId(Long userId);
 
     /**
      * 获取购物车中未勾选的商品数
@@ -28,9 +28,9 @@ public interface CartMapper {
      * @param userId 用户id
      * @return 未勾选的商品数
      */
-    int countUnCheckedByUserId(Integer userId);
+    int countUnCheckedByUserId(Long userId);
 
-    int deleteByUserIdAndProductIds(@Param("userId") Integer userId, @Param("ids") List<Integer> ids);
+    int deleteByUserIdAndProductIds(@Param("userId") Long userId, @Param("ids") List<Long> ids);
 
     /**
      * 设置用户购物车全选/全不选
@@ -38,9 +38,9 @@ public interface CartMapper {
      * @param userId  用户id
      * @param checked 全选/全不选
      */
-    void updateAllCheckedStatusByUserId(@Param("userId") Integer userId, @Param("checked") int checked);
+    void updateAllCheckedStatusByUserId(@Param("userId") Long userId, @Param("checked") int checked);
 
-    void updateCheckedStatusByUserIdAndProductId(@Param("userId") Integer userId, @Param("productId") Integer productId, @Param("checked") int checked);
+    void updateCheckedStatusByUserIdAndProductId(@Param("userId") Long userId, @Param("productId") Long productId, @Param("checked") int checked);
 
     /**
      * 获取指定用户购物车中商品总数量
@@ -48,7 +48,7 @@ public interface CartMapper {
      * @param userId 用户id
      * @return 商品总数量，用户id不存在返回0
      */
-    int countCartProductByUserId(Integer userId);
+    int countCartProductByUserId(Long userId);
 
     /**
      * 获取指定用户勾选的商品列表
@@ -56,5 +56,5 @@ public interface CartMapper {
      * @param userId 用户id
      * @return 购物车商品列表
      */
-    List<Cart> listCheckedCartByUserId(Integer userId);
+    List<Cart> listCheckedCartByUserId(Long userId);
 }
