+++ src/main/java/cn/waynechu/mmall/mapper/CategoryMapper.java
@@ -5,17 +5,17 @@ import cn.waynechu.mmall.entity.Category;
 import java.util.List;
 
 public interface CategoryMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Category record);
 
     int insertSelective(Category record);
 
-    Category selectByPrimaryKey(Integer id);
+    Category selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Category record);
 
     int updateByPrimaryKey(Category record);
 
-    List<Category> getChildrenParallelCategory(Integer parentId);
+    List<Category> getChildrenParallelCategory(Long parentId);
 }
